Extracted out the SocketInterfaceSwap class.
